Reorganized namespaces
Logger is apart of Gummi.Core. Scripts in the core may refer to each other, but should never refer to scripts outside of the core. Scripts outside of the core may be dependent on the core, though.

Command pattern business has been moved to Gummi.Pattern space. Base coding patterns should be here.